requirements.txt
langchain==0.0.271
openai==0.27.9
tiktoken==0.4.0
slack-bolt==1.18.0
supabase==1.0.4
pinecone-client==2.2.2
pypdf==3.15.2
python-dotenv==1.0.0